+++ src/App.css
@@ -10,7 +10,7 @@
   padding: 0.5em 0.5em;
   margin: 1em 1em;
   width: 20em;
-  height: 30em;
+  height: 35em;
   object-fit: cover;
   border-radius: 10px;
 }
@@ -28,6 +28,27 @@
   width: 400px;
   height: 100px;
   margin: 10px 10px;
+}
+
+.header-container {
+  display: flex;
+  width: 100%;
+  height: 200px;
+  justify-content: space-between;
+}
+
+.title-container {
+  flex: 0 0 25%;
+}
+
+.link-container {
+  flex: 0 0 75%;
+  display: inline-flex;
+  justify-content: space-between;
+}
+
+.header-link {
+  margin: auto;
 }
 
 ul {
@@ -51,6 +72,14 @@ img {
   object-fit: cover;
   border-radius: 10px;
   box-shadow: 10px 10px hotpink;
+}
+
+a {
+  border: 0px;
+  padding: 5px 5px;
+  text-align: center;
+  text-decoration: none;
+  display: inline-block;
 }
 
 #comment-success-text {
