+++ src/components/ArticlePage.jsx
@@ -1,0 +1,8 @@
+export const ArticlePage = () => {
+    return (
+        <>
+            <Title/>
+            <ArticleList/>
+        </>
+    )
+}
