+++ src/components/Header.jsx
@@ -2,18 +2,22 @@ import { Link } from "react-router-dom";
 
 export const Header = () => {
     return (
-      <header className="header">
-          <nav>
-            <Link to={"/"}>
+      <header className="header-container">
+          <div className="title-container">
+            <h1>NC News</h1>
+            <p>Your daily dose of knowledge</p>
+          </div>
+          <nav className="link-container">
+            <Link to={"/"} className="header-link">
               <p>Home</p>
             </Link>
-            <Link to={"/ncnews/users"}>
+            <Link to={"/ncnews/users"} className="header-link">
               <p>Users</p>
             </Link>
-            <p>Login</p>
+            <p className="header-link">Login</p>
           </nav>
-          <h1>NC News</h1>
-          <p>Your daily dose of knowledge</p>
       </header>
     );
   };
+
+
