+++ src/components/Title.jsx
@@ -1,0 +1,1 @@
+export const Title = () => <h2 className="articles-list-title">Some news</h2>
